+++ onnx-opl/src/lib.rs
@@ -12,6 +12,7 @@ pub mod is_inf;
 pub mod is_nan;
 pub mod lrn;
 pub mod ml;
+pub mod non_max_suppression;
 
 pub trait WithOnnx {
     fn with_onnx(self) -> Self;
@@ -28,6 +29,7 @@ impl WithOnnx for tract_nnef::framework::Nnef {
 fn onnx_opl_registry() -> Registry {
     let mut registry: Registry = Registry::new("tract_onnx");
     ml::register(&mut registry);
+    non_max_suppression::register(&mut registry);
     registry.register_unit_element_wise("tract_onnx_erf", &erf::Erf {});
     registry.register_element_wise(
         "tract_onnx_isinf",
